+++ src/cpp/json_parser.h
@@ -5,16 +5,16 @@
 #include <vector>
 #include <unordered_map>
 
-// Parse product from JSON
+// Parse product from JSON string
 Product parseProductJson(const std::string &productJson);
 
-// Parse substances from JSON
+// Parse substances from JSON string
 std::vector<Substance> parseSubstancesJson(const std::string &substancesJson);
 
-// Parse effect multipliers from JSON
-std::unordered_map<std::string, double> parseEffectMultipliersJson(const std::string &effectMultipliersJson);
+// Parse effect multipliers from JSON string - returns integers multiplied by 100
+std::unordered_map<std::string, int> parseEffectMultipliersJson(const std::string &effectMultipliersJson);
 
-// Parse and apply substance rules from JSON
+// Parse substance rules from JSON string and apply to substances
 void applySubstanceRulesJson(
     std::vector<Substance> &substances,
     const std::string &substanceRulesJson);
